feat(call): custom SIP headers on outgoing INVITE; audio/video conf fixes
- Thread optional headers map through placeCall/placeCallWithMedia ->
  Manager::outgoingCall/newOutgoingCall -> SIPAccount, injected into INVITE
  via pjsip_generic_string_hdr in SIPStartCall.
- Manager::detachLocalParticipant: no-arg API now resolves and detaches the
  active conference, releasing the local camera VideoInput ref after logout.
- Manager::addAudio: pin PLAYBACK+CAPTURE device streams for conference
  participants so re-invites don't tear down mic/speaker on WASAPI/PortAudio.
- VideoMixer::render_frame: use floating-point source aspect for black-border
  crop selection to avoid integer-truncation mis-crop.

+++ src/media/video/video_mixer.cpp
@@ -1226,8 +1226,12 @@ VideoMixer::render_frame(VideoFrame& output,
         // calculate width and height for cropping
         int width = 0, height = 0;
         if (remove_black_borders_ && not source->isBig) {
-            // calculte cropping according to aspects
-            if (grid_aspect_ > input->width() / input->height()) {
+            // calculte cropping according to aspects.
+            // Use floating-point source aspect: integer input->width()/input->height()
+            // truncates (e.g. 16:9 -> 1), which mis-selects the crop branch and can
+            // compute a crop taller than the input for sources whose true aspect
+            // sits between grid_aspect_ and 2.0.
+            if (grid_aspect_ > static_cast<double>(input->width()) / input->height()) {
                 width = input->width();
                 height = width / grid_aspect_;
             } else {
